change userid type to uint64

+++ internal/adapter/storage/postgres/repository/permission.go
@@ -21,7 +21,7 @@ func NewPermissionRepository(log logger.Logger, db *sql.DB) *PermissionRepositor
 	}
 }
 
-func (r PermissionRepository) GetUserPermissionKeys(ctx context.Context, userID uint) ([]domain.PermissionKeyType, error) {
+func (r PermissionRepository) GetUserPermissionKeys(ctx context.Context, userID uint64) ([]domain.PermissionKeyType, error) {
 	rows, err := r.db.QueryContext(
 		ctx,
 		`SELECT DISTINCT p.key
